+++ packages/types/src/params/msg.rs
@@ -128,7 +128,7 @@ pub struct ManagedVaultConfigResponse {
     /// Minimum creation fee in uusd for managed vaults
     pub min_creation_fee_in_uusd: u128,
     /// List of code ids for managed vaults
-    pub code_ids: Vec<u32>,
+    pub code_ids: Vec<u64>,
 }
 
 #[cw_serde]
@@ -189,8 +189,8 @@ pub enum EmergencyUpdate {
 
 #[cw_serde]
 pub enum ManagedVaultConfigUpdate {
-    AddCodeId(u32),
-    RemoveCodeId(u32),
+    AddCodeId(u64),
+    RemoveCodeId(u64),
     SetMinCreationFeeInUusd(u128),
 }
